Swig: Explicit namespace to exported function
C++11 provides a std::rank function that conflicts with the recently
added smil::rank morphological filter, especially when building the
Python wrapper.

One way to circumvent this issue, as the current commit shows it, is
to ensure that Swig exported SMIL functions are prefixed by their
namespace.
This works but has not been extensively tested.

Another solution is to rename the smil::rank function. This
nevertheless does not treat potential future conflicts with other
libraries.