CMake/Preinstall/pkg-config: Fix link to dependencies
When building Smil as static archive (the default), put cpp
dependencies into the Libs field in the pkg-config file, so that Smil
users will find them for linking.

When building Smil as a dynamic shared object (by setting
-DBUILD_SHARED_LIBS), keep cpp dependencies into the Libs.private
field. In this way, applications that statically link with Smil .so
will also link to them (if I understand correctly the semantics of the
Libs.private field...).