Base/test_draw: Search DejaVu font in several places
This test hardcodes the path of the DejaVu font used for drawing, but
the path used is OS dependant and does not work at least on ArchLinux.

To be backward-compatible with C++<11, I used C arrays and the
ifstream C++ class, instead of the C++11 std::array and the new C++17
std::filesystem APIs.