IO/DImageIO_PNG: Rewrite template specializations for MSVC
MSVC could not find the right specializations for the
PNGImageFileHandler<UINT8>::read and write functions in spite of the
declarations of these functions in the header file, and instead linked
with the virtual, non-specialized PNGImageFileHandler<T>::read and
write functions.

To fix this, we have to explicitely declare the template
specializations of the PNGImageFileHandler<T> and declare the relevant
read/write implementations in these classes. Which was done here with
the UINT8, UINT16 and RGB types.

As a consequence, the specialized functions definitions in
DImageIO_PNG.cpp should drop the template<> in front of their
signature.

This new implementation should be compatible with GCC, Clang and
MSVC, and without having to rely on weird template specialization
declarations, such as the one hidden behind the TEMPL_SPEC_DECL macro.